Fix app router error.
